ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
openai==1.58.1
httpx==0.28.1
python-multipart==0.0.20
structlog==24.4.0
pytesseract==0.3.13
Pillow==11.1.0
openpyxl==3.1.5
python-docx==1.1.2
pdf2image==1.17.0
PyMuPDF==1.25.3
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
